Markdown publication now handles multiplication symbol properly
In recent releases of Matlab (R2016b+, I think), Matlab prints to the
command window using html symbols rather than an x symbol when showing
array sizes.  These symbols are now replaced with a simple x when
publishing to markdown (note that this is done as a text replacement
in publishreadme.m, not in the XSL stylesheet itself; eventually I should
add a more robust fix in the stylesheet).